requirements.txt
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
streamlit==1.29.0
openai==1.35.3
python-dotenv==1.0.0